Move popup component along with styling into separate folder.
