+++ frontend/political_scale/src/popUp/PopUp.styled.js
